lib/env_dpdk: update PCI headers for DPDK 22.11.4
Just this week DPDK 22.11.4 was released, which
included a change in the pci_dpdk API.

The change is non-breaking for SPDK use of the API.
It was introduced as backport from DPDK 23.07.

If rte_pci_pasid_set_state() is to be used in future,
a new copy and implementation should be created.
To catch the case of someone unaware of this versioning,
a redefinition of this function was added.
It should have been present since SPDK supported DPDK 23.07.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: If591d35d8d21bb3f5e734dc9fd9ccf4b05cc7513
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21578
Reviewed-by: Michal Berger <michal.berger@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot