module/raid: DIX support for RAID 0
Added an ability for the application to use DIX protection
(separate metadata) with the RAID bdev, with RAID level 0.

In order to handle the protection information correctly for
RAID bdevs, the patch introduces Reference Tag remapping between
the RAID bdev and base bdevs on read and write paths.

Change-Id: I4056a3e4410a4df3282b993b655f6e6e2715e5a9
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21618
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>