lib/iscsi: Add RPC method iscsi_get_stats
In a production environment, the iscsi_get_connections RPC method is
commonly used to monitor the number of SPDK connections. It is often
necessary to issue an alarm when there are too many iSCSI connections.
By default, the maximum number of connections in SPDK is defined by
MAX_ISCSI_CONNECTIONS as 1024.

However, the return value of the iscsi_get_connections method contains
a significant amount of information. In certain cases, only the number
of connections is required, and the overhead of parsing JSON becomes
a considerable burden, especially when there are numerous connections.

To address this issue, the rpc method iscsi_get_stats has been added.
When this method is used, only the count of connections in each state
will be provided, eliminating the need for detailed information parsing.

Change-Id: I45f4a4ebb2d4a5a6fe9eee28a3cbeb4e05009b56
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>
Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21629
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>