nvme/auth: support DH-HMAC-CHAP key transformation
It is now possible to use transformed key by running an HMAC over the
configured DH-HMAC-CHAP key and the NQN of the host.  As described in
the NVMe spec, this mode is automatically selected when the configured
key specifies non-zero hash function (i.e. DHHC-1:{01,02,03} selects
HMAC with sha256, sha384, and sha512 respectively).

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ia24cbffe0485cf9f55c56c14c9e6505e1a5585ba
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22016
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>