configure: Hint compiler at what linker to use via -fuse-ld
When user provides custom $LD, which points at something different than
the bfd linker, it's not considered while testing the dependencies via
BUILD_CMD[@]. Granted, the proper linker is then set via our mk infra,
but the same way as BUILD_CMD[@] uses $CC, ${C,LD}FLAGS it should also
consider other relevant opts.

There's a small conundrum that forces as to do the same thing for the
isa-l* - in case user tries to provide extra params to the linker via
LDFLAGS and which may be specific to a given $LD_TYPE, the isa-l*
configure will fail while trying to compile its dependencies. This is
because it still slurps the LDFLAGS but the compiler keeps using the
default linker. As we overwrite isa-l*'s CFLAGS via its cmdline we
can't pass that information through the environment as well. So as a
best effort, without dancing too much, just pass -fuse-ld to it based
on what detect_cc.sh already provided. From perspective of the actual
isa-l* build this is still foobar, since this param won't be actually
used (and clang won't mind telling us about it) so to silence any
warnings about it slap no-unused-command-line-argument as well.

Change-Id: If392c64099b0b50c879d504151763723924edf77
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21769
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot