nvme: free process structures for all transports
Fixed memory leak in nvme_ctrlr_add_process function.
For TCP and RDMA transports, memory allocated for ctrl_proc
structure was never freed, as the nvme_ctrlr_free_processes()
function was not called at the end of controller destruction.

Now the memory leak is fixed, and a call to
nvme_ctrlr_free_processes() is added to the
nvme_ctrlr_destruct_finish() function, so it is called
for all types of transports.

Also the stubs for nvme_ctrlr_free_processes() are removed
from unit tests, as they are not needed now.

Suggested-by: Jim Harris <jim.harris@samsung.com>

Change-Id: I54a2bdbb32b4f1c56d9c2fccfb1e93cb932f375d
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21782
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>