app/spdk_dd: use configure option to build with uring
The support for uring in spdk_dd app was not actually compiled
in as the relevant configure option went unused.

This was not caught as there is no way to force spdk_dd
to use uring, and relevant test function to check the
linking was broken.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: If744ada604c270975b1814c3d4a7d3668a4cc8b9
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21836
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot