trace: add size field to spdk_trace_file
This is more straightforward than using an extra entry in the
lcore_history_offsets array to store the file size. This will also make
more sense if/when we add other types of variable sized data to the trace
files in the future.

The bulk of the changes here are in the trace_record app. We now need to
write out the file_size explicitly to the out_fd.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I3f22359b23ae5915146142b76e38924568782334
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21791
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>