test/dd: check spdk_dd direct link to liburing
If SPDK is compiled with uring support for bdev or sock
modules, this does NOT have to mean that spdk_dd was too.

check_liburing() was using ld, that will list indirect
dependencies, rather than only verifying those directly used.
Despite spdk_dd not actually compiling with liburing,
this function didn't catch that.

This could be resolved by passing empty LD_LIBRARY_PATH,
which would make ld unable to resolve those dependencies.

Instead of that workaround, objdump is used to verify only
directly linked libraries.
Removed the check if liburing actually exists, since objdump
does not point to their actual locations. If it is missing,
that will be caught when running the test anyway.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I13edc36d9667732c170eecccdcb3b5f334b87b61
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21835
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>