bdev: check separate md_buf and accel_sequence in IO submission
The bdev layer doesn't support separate metadata buffers for IOs with an
accel sequence.  Previously, we wouldn't allow registering a bdev with
separate metadata buffers that advertised support for accel sequences.
However, this approach has a couple of issues: firstly, it prevents users
from creating such bdevs even if they don't intend to use accel, and
secondly, it still allows users to send IOs with an accel sequence to
bdevs with separate metadata if they didn't support accel.

To fix this, the check from bdev_register() was replaced with checks in
the submission functions.

Reported-by: Karl Bonde Torp <k.torp@samsung.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ie39dcbf2e5454d0d04ef57eedf29eefc7be7a899
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21859
Reviewed-by: Karl Bonde Torp <k.torp@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>