build: ensure env deps are included when linking
The dependencies of the env library were not included when linking apps.
Right now, it isn't a big problem for env_dpdk, as it only depends on
log and util, which almost all other libraries also depend on, but it
has to be fixed if we ever needed to add more dependencies.

Reported-by: Jacek Kalwas <jacek.kalwas@intel.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I482845f74598d5525dc16d5e69e85db7fc4424f6
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21873
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@intel.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>