fuzzer: enable reporting address of leaked memory
This patch turns on reporting of the address of leaked
memory object, when a leak sanitizer is detecting a memory
leak during fuzz tests. This is done by setting
LSAN_OPTIONS vaiable to 'report_objects=1'

Can be used to verify if the leaked object is really
a leak, or during memory leak debugging.
Example output:

  Objects leaked above:
  0x60700000a820 (72 bytes)

  SUMMARY: AddressSanitizer: 72 byte(s) leaked in 1 allocation(s).

If 'report_objects' is not enabled, the address of leaked object
is not provided (only the size is reported)

Change-Id: I0d59ff82096b64ed17fb89e369b439c9f565f500
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21896
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>