app/event: introduce raw json data app opt
Right now, if application wants to create some JSON configuration at
run time, it would need to create and manage some temporary file,
which may be not desired.

To avoid it, introduce 2 new app options (raw json data buffer and its
size), similarly as it was recently done for spdk_subsystem_load_config.

This way of providing configuration is going to be mutually exclusive
with existing file based approach, so that SPDK app framework is not
required to resolve any conflicts between these two and it is left for
application developers.

Signed-off-by: Maciej Szulik <maciej.szulik@intel.com>
Change-Id: Id6d5a560309863e317ac99db1a08917edd75256e
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22212
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@intel.com>