test/nvmf/fips: Don't export OPENSSL_FORCE_FIPS_MODE
This is redundant since we already do this explicitly in the custom
.cnf. Also, it breaks initialization under redhat's openssl 3.1.1 where
the 0032-Force-fips.patch changed quite significantly - it seems it no
longer checks if given provider was already activated|loaded hence the
provider_conf_activate() fails when the second attempt is made.

Previous iteration of this patch was flipping ok to 1 when provider was
already active here it remains 0.

To avoid this, simply don't export OPENSSL_FORCE_FIPS_MODE to make sure
all needed providers (base + fips) are loaded solely based on our .cnf.

Change-Id: I00e007360fac51754fa8c246394e4a5ecf4ffc07
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21972
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>