sock: add trace points for sock requests
Add tracepoints for queue/pend/complete for each sock request.

Note that we don't want to calculate the total request length when the
SOCK_REQ_QUEUE tracepoint is not enabled, so we use the new spdk_trace_tpoint_enabled()
macro to do that calculation only when necessary.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I54d112cd0926c2e683968b17ba05b2da97061355
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22099
Reviewed-by: Seung yeon Shin <syeon.shin@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>