trace: add spdk_trace_tpoint_enabled()
This macro allows callers to restrict any extra prep code from executing when
its associate tpoint is not enabled.

This can be useful for certain tracepoints that need to calculate some value for
the trace point - for example, add up the lengths of an iov array. We will
use it for exactly that reason in an upcoming patch.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: If1c3f1a6801aeb62ae4bcb763838d6101464dab9
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22098
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>