trace_parser: add flag for arguments used for related objects
Some tracepoints have an argument that is a context pointer for some other
object type. The trace app already decode this information, showing the ID for
that related object.

Since the ID is already shown, the context pointer itself isn't useful, so
add a flag to note those arguments, so we can omit printing the context
pointer.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I3f6521ff5594cdc4c635c23a612c8f0d2f3717dd
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22102
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>