spdk_trace: default on Linux to newest trace file in /dev/shm
It is kind of a pain to have to *always* have to specify the filename
or app parameters to spdk_trace, especially when you have the app currently
running or the app just exited.

So use ntfw() to walk the files in /dev/shm when -f and -s are not specified,
and use the newest file (assuming one is found).

A few caveats:
1) This only works on Linux, mainly because FreeBSD does not have a /dev/shm
   mount for shm files.
2) Search /dev/shm is not foolproof, since we don't know for sure whether a
   file there is an SPDK trace file. But looking for the "_trace." substring
   should be enough. It should guard against trying to use other non-trace
   files.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: If130e6f5918a228b6f9d4576e16d37e05bfc4995
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22104
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>