nvmf/vfio_user: reject fabrics commands
Currently these are rejected, because spdk_nvmf_request_exec() will only
allow fabrics transports to execute fabrics commands.  Since we'll want
to remove that restriction, vfio-user needs to explicitly check for this
opcode.

Additionally, this patch also changes the status code for the
reservation commands to INVALID_OPCODE, as this is what should be done
when reservations aren't supported by a controller.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I00b530db7f6f505e926fa601a007cad043f5ed80
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22236
Reviewed-by: Thanos Makatos <thanos.makatos@nutanix.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jacek Kalwas <jacek.kalwas@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>