nvmf: allow commands depending on qpair state
The state of a qpair is now used to determine whether a command is
allowed to execute.  This replaces checks testing qpair->ctrlr == NULL
to see if the CONNECT command has been received.  Relying on the state
will make it easy to add checks for more states (e.g. authentication) in
the future.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I49274f3525d4f2fc267cef32a9c68824ea5d4705
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22589
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>