util: add SPDK_GET_FIELD()
The sizes of many public structures are tracked in order to ensure ABI
compatibility.  This requires checking the size of a structure before
accessing any of its fields.

This patch introduces a macro that makes this a bit more convenient and
uses it to get fields from spdk_bdev_ext_io_opts.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Id0e92d2df02cfa24fda148353ae049d7fc7183ea
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22591
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot