bdev/nvme: handle uuid generation errors
This is second patch in series that will implement
error code handling during uuid generation.

The nvme_generate_uuid() function definition is changed,
it takes additional parameter as a pointer to
location where the uuid is stored, and returns error
code instead of the uuid structure.

If in nvme_disk_create() function there is an
error code returned from nvme_generate_uuid(),
an error message is logged and the functions returns
with the error code passed from nvme_generate_uuid().

Now, when we are handlig errors from nvme_generate_uuid(),
an assert that was verifying status of snprintf()
operation is replaced by regular 'if' and returning
error code in case of fail.

Change-Id: Iaf1a76597d9aa98c433ebb9d69ba6f9a22773deb
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22291
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>