test/iscsi: delete fio_remote_nvme test
This test is never executed in continous
integration pipelines. It needs both
SPDK_TEST_ISCSI and SPDK_TEST_NVMF flags
to be run which would result in a very
long execution time (up to 1 hour).
It is unlikely that such combination
be ever used in CI for testing.

Furthermore - the whole scenario of using
SPDK NVMe-oF Target & Initiator & iSCSI
Target in a single script seems a bit
convoluted. Unfortunately the change
introducing the test did not provide any
description if this is a real use case.
Pretty sure all of actions done here are
already covered by other test scripts
separately.

Change-Id: I9042d50a68a83295dd5feeb89617184ccd18b565
Signed-off-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22267
Reviewed-by: Michal Berger <michal.berger@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>