identify: fix zone report bufsize calculation
It used to solely rely on the max_xfer_size of a controller, which, on
some transports (e.g. TCP), can be unlimited and is set to UINT32_MAX.
We obviously don't need to allocate that much memory, so limit it to the
number of zones we're trying to show.

Fixes #3297.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I2a54cc5d63234045733d5c848a38ac8957a71033
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22323
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>