nvmf: method for getting DH-HMAC-CHAP keys
This is a helper function that returns host's key in a thread-safe
manner.  To ensure that a key object remains valid, even if the host is
removed immediately after calling this function, it's duplicated and
users are required to release it with spdk_keyring_put_key().

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I2b56c1835c119b0dcf134ce45d93792710bd1cf0
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22667
Reviewed-by: Seung yeon Shin <syeon.shin@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>