nvmf/auth: execute DH-HMAC-CHAP_reply message
The reply message is executed by calculating the response to the
challenge and comparing it from the response received from the host.

Initially, bidirectional authentication won't be supported and a message
with the cvalid bit set is treated as an error.  The code to enable that
will be added later on, in a separate commit.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Icf3ea216dcddec66e768485f894cb51807313ffb
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22668
Reviewed-by: Seung yeon Shin <syeon.shin@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot