trace: increase SPDK_TRACE_MAX_LCORE to 1024
We limited this to 128 originally since that was aligned with the 128 max
lcores supported by DPDK on x86.

But 128 isn't a suitable maximum anymore. SPDK trace files are also dynamic
now, meaning we don't allocate trace buffers for unused lcores. There is a
small amount of space that we allocate per-lcore for thread names, but this
pales in comparison to the size of a typical trace buffer.

So bump the number to 1024. This will allocate an extra 14KB of data per
trace file ((1024 - 128) * 16B) for thread names, but each core's trace
buffer by default is 1MB so this seems like a minor amount of additional
overhead.

Fixes issue #3301.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I87e656fc7c8b2002e90d71f6eb11fb4699582f63
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22418
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>