bdev/raid1: Support resize when increasing the size of base bdevs
Implement the resize function for RAID1. raid1_resize() calculate the
new raid_bdev's block count and if it is different from the old block
count, call spdk_bdev_notify_blockcnt_change() with the new block count.

Add a simple functional test for this feature. The test is to create
a raid1 bdev with two null bdevs, resize one null bdev, check if the
raid1 bdev is not resize, resize another null bdev, check if the raid1
bdev is resized.

The commit is an overall copy/adaptation of the work of Shuhei Matsumoto
with raid0.

Change-Id: I0737ba3fe4e1260d35d51e7b022c9cfd1eb6db58
Signed-off-by: Loïc Yavercovski <loic.yavercovski@gmail.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22619
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>