test/app/stub: add command line option to set default io_queue_size
Commit ebdb06b ("stub: fix a io_queue_size limitation") set the default
ctrlr io_queue_size to UINT16_MAX. Details from the commit message and
old GerritHub URL aren't clear what the performance issue was though.
SPDK performance reports are all generated without using stub app to my
knowledge.

This causes problems when using the stub app with performance tools such as
fio. fio doesn't specify io_qpair options, so it will just default to whatever
was set when the controller was attached. Without the stub, it will use
default (256) but with the stub running it uses whatever max supported
by the controller (MQES). This can cause performance anomalies, but less
pleasant are the out-of-memory issues due to all of the extra trackers this
tries to allocate for tests with lots of jobs.

So change this up so that by default, we don't change the default - only
change it if user passed some value via command line.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ia68a76b6272b2ef307d6f4e346e204d80c474828
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22448
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>