test/app/stub: enforce stub to only run as a primary process
The stub process is only meant to be run as a primary process, it is not useful
at all as a secondary process. We can use spdk_app_opts::env_context
to specify to DPDK that the process should be initialized as a primary process.
That way if a stub process is already running at the same shm_id it will exit
with error immediately.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I9091ebc4b996c55531e82fc6231aebb210e5810b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22476
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>